+++ darglint/utils.py
@@ -77,16 +77,6 @@ class AstNodeUtils(object):
         dot.append('}')
         return '\n'.join(dot)
 
-# def generic_visit(self, node):
-#         """Called if no explicit visitor function exists for a node."""
-#         for field, value in iter_fields(node):
-#             if isinstance(value, list):
-#                 for item in value:
-#                     if isinstance(item, AST):
-#                         self.visit(item)
-#             elif isinstance(value, AST):
-#                 self.visit(value)
-
 
 class CykNodeUtils(object):