+++ darglint/flake8_entry.py
@@ -18,7 +18,7 @@ from .config import (
 )
 
 
-__version__ = '1.5.3'
+__version__ = '1.5.4'
 
 
 class DarglintChecker(object):